+++ CloudRemoting.psm1
@@ -1,4 +1,4 @@
-Write-Verbose "EC2Remoting module"
+Write-Verbose "CloudRemoting module"
 
 $functionFilter = Join-Path $PSScriptRoot "Functions\*.ps1"
 Get-ChildItem -Path $functionFilter -Recurse | Foreach-Object {