+++ Functions/Common.ps1
@@ -1,0 +1,3 @@
+# Common Aliases
+New-Alias -Name rdp -Value Enter-RdpSession -Force
+New-Alias -Name ec2rdp -Value Enter-EC2RdpSession -Force